pyproject.toml
[tool.poetry]
name = "fair-funding"
version = "0.1.0"
description = "Fair Funding Platform"
authors = ["Unstoppable <team@unstoppable.ooo>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
vyper = "^0.3.7"


[tool.poetry.group.dev.dependencies]
hypothesis = "^6.67.1"
titanoboa = {git = "https://github.com/vyperlang/titanoboa"}
pytest-watch = "^4.2.0"
blackadder = "^0.1.1"
black = "^22.10.0"
coverage = "^6.5.0"
pytest-only = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = [ "ignore::DeprecationWarning" ]
